the i3 module is basically obsolete at this point
aside from making the i3 json protocol the default mode for this thing,
i've also moved to sway, so the whole i3 ipc thing isn't really working
out anymore. and, if you use i3-bar or swaybar, the workspaces are
displayed anyways.

so, lemonbar and i3 are out, sway and built-in workspace numbers are in
:) i will most likely choose to support lemonbar and the i3 module once
again (in which case.... uhhhh... i'll need to get the i3 module
back...)