switching over to a daemon and client system
clients can send messages to the daemon for status updates, so that
"rapidfire checking" doesn't have to happen anymore. This was causing
significant performance issues for the system that muse-status was
running on. Now, without rapidfire, the system can run fine without
overworking to check for volume, brightness, bspwm, and window updates